gui/requirements.txt
cffi==1.17.1
cryptography==45.0.5
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyYAML==6.0.2
